requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-admin==2.0.0
django-allauth==0.41.0
django-debug-toolbar==2.1
django-excel-response2==2.0.8
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.11.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
screen==1.0.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
xlwt==1.3.0
